+++ fishyrl/environments.py
@@ -1,6 +1,7 @@
 """Environment definitions for FishyRL."""
 
 import enum
+import warnings
 from abc import abstractmethod
 from typing import Any
 
@@ -8,6 +9,21 @@ import gymnasium as gym
 import numpy as np
 
 from . import utilities as frl_utilities
+
+# Try importing RLGym
+# TODO: Maybe move import inside VectorizedRLGymEnvironment
+try:
+    import rlgym.api as rlapi
+    import rlgym.rocket_league.action_parsers as rlact
+    import rlgym.rocket_league.common_values as rlcommon
+    import rlgym.rocket_league.done_conditions as rldone
+    import rlgym.rocket_league.obs_builders as rlobs
+    import rlgym.rocket_league.reward_functions as rlreward
+    import rlgym.rocket_league.rlviser as rlviser
+    import rlgym.rocket_league.sim as rlsim
+    import rlgym.rocket_league.state_mutators as rlstate
+except ImportError:
+    pass
 
 
 class VectorizedEnvironment:
@@ -217,6 +233,259 @@ class VectorizedGymEnvironment(VectorizedEnvironment):
         return VectorizedGymEnvironment(**new_kwargs)
 
 
+class VectorizedRLGymEnvironment(VectorizedEnvironment):
+    """A vectorized environment wrapper for RLGym environments."""
+    def __init__(self, num_envs: int = 1, allow_rendering: bool = False, **init_kwargs: dict[str, Any]) -> None:
+        """Initialize the vectorized RLGym environment.
+
+        :param num_envs: The number of parallel environments to create. (Default: ``1``)
+        :type num_envs: int
+        :param allow_rendering: Whether to allow rendering of the environment, currently unsupported. (Default: ``False``)
+        :type allow_rendering: bool
+        :param init_kwargs: Additional keyword arguments for initializing the RLGym environment. (Default: ``{}``)
+        :type init_kwargs: dict[str, Any]
+
+        """
+        # Parameters
+        self._num_envs = num_envs
+        self._allow_rendering = allow_rendering
+        self._init_kwargs = init_kwargs
+
+        # Initialize environments
+        self._envs = []
+        for _ in range(self._num_envs):
+            # Actions
+            action_parser = rlact.RepeatAction(
+                rlact.LookupTableAction(), repeats=8)
+            # Truncation and termination
+            termination_cond = rldone.GoalCondition()
+            truncation_cond = rldone.AnyCondition(
+                rldone.NoTouchTimeoutCondition(30), rldone.TimeoutCondition(300))
+            # Rewards
+            reward_fn = rlreward.CombinedReward(
+                (rlreward.GoalReward(), 100), (rlreward.TouchReward(), 1))
+            # Observations
+            obs_builder = rlobs.DefaultObs(
+                zero_padding=None,
+                pos_coef=np.asarray([1 / rlcommon.SIDE_WALL_X, 1 / rlcommon.BACK_NET_Y, 1 / rlcommon.CEILING_Z]),
+                ang_coef=1 / np.pi,
+                lin_vel_coef=1 / rlcommon.CAR_MAX_SPEED,
+                ang_vel_coef=1 / rlcommon.CAR_MAX_ANG_VEL,
+                boost_coef=1 / 100.)
+            # Training conditions
+            state_mutator = rlstate.MutatorSequence(
+                rlstate.FixedTeamSizeMutator(2, 2),
+                rlstate.KickoffMutator())
+            # Create environment
+            env = rlapi.RLGym(
+                action_parser=action_parser,
+                termination_cond=termination_cond,
+                truncation_cond=truncation_cond,
+                reward_fn=reward_fn,
+                obs_builder=obs_builder,
+                state_mutator=state_mutator,
+                transition_engine=rlsim.RocketSimEngine(),
+                renderer=rlviser.RLViserRenderer(tick_rate=120. / action_parser.repeats) if self._allow_rendering else None)
+            # Record
+            self._envs.append(env)
+
+        # Call an initial reset to initialize observations
+        self.reset()
+
+    @property
+    def num_envs(self) -> int:
+        """The number of environments.
+
+        :type: int
+
+        """
+        return self._num_envs
+
+    @property
+    def render_fps(self) -> int:
+        """The number of frames per second while rendering.
+
+        :type: int
+
+        """
+        # return rlcommon.TICKS_PER_SECOND / self._envs[0].action_parser.repeats  # 1 frame per action
+        # return rlcommon.TICKS_PER_SECOND  # Actual ticks
+        return self._envs[0].renderer.tick_rate
+
+    @property
+    def obs_shape(self) -> np.ndarray:
+        """The shape of the observation space.
+
+        :type: np.ndarray
+
+        """
+        return list(self._envs[0].observation_spaces.values())[0][1]
+
+    def _construct_actions(self, actions: np.ndarray) -> list[dict[str, np.ndarray]]:
+        """Construct a list of dicts mapping agents to one-hot actions from an array of discrete actions.
+
+        :param actions: An array of discrete actions, with shape (num_envs, num_agents).
+        :type actions: np.ndarray
+        :return: A list of dicts mapping agents to one-hot actions, with length num_envs.
+        :rtype: list[dict[str, np.ndarray]]
+
+        """
+        # Get action space
+        # NOTE: Assumes all environments have the same action space, which should be true for this implementation
+        action_space_type, action_space_size = list(self._envs[0].action_spaces.values())[0]
+
+        # Check that action space is correct
+        if action_space_type != 'discrete':
+            raise NotImplementedError("Only discrete action spaces are currently supported.")
+
+        # Decompose array into list of dicts mapping agents to one-hot actions
+        new_actions = []
+        for env, env_actions in zip(self._envs, actions):
+            # Get agents
+            agents = env.agents
+
+            # Create dict of actions for each agent, converting from discrete to one-hot
+            env_new_actions = {agent: np.zeros(action_space_size, dtype=bool) for agent in agents}
+            for i, agent in enumerate(agents):
+                env_new_actions[agent][env_actions[i]] = 1
+            new_actions.append(env_new_actions)
+
+        return new_actions
+
+    def action_sample(self) -> np.ndarray:
+        """Sample an action from the action space.
+
+        :return: An array of sampled actions, with shape (num_envs, num_agents).
+        :rtype: np.ndarray
+
+        """
+        # Get action space
+        action_space_type, action_space_size = list(self._envs[0].action_spaces.values())[0]
+
+        # Check that action space is correct
+        if action_space_type != 'discrete':
+            raise NotImplementedError("Only discrete action spaces are currently supported.")
+
+        # Sample actions
+        actions = []
+        for env in self._envs:
+            # Randomly sample one action per agent
+            env_actions = np.random.randint(action_space_size, size=len(env.agents))
+            actions.append(env_actions)
+
+        # Stack actions
+        actions = np.stack(actions, axis=0)
+
+        return actions
+
+    def reset(self, seed: int = None, **kwargs: dict[str, Any]) -> tuple[np.ndarray, dict[str, Any]]:
+        """Reset the environments.
+
+        :param seed: An optional seed for the environments, currently unsupported. (Default: ``None``)
+        :type seed: int, optional
+        :param kwargs: Additional keyword arguments to pass to the environment reset method.
+        :type kwargs: dict[str, Any]
+        :return: A tuple containing the initial observations and an empty info dictionary.
+        :rtype: tuple[np.ndarray, dict[str, Any]]
+
+        """
+        # Raise warning about ignoring seed
+        if seed is not None:
+            warnings.warn("RLGym environments do not currently support seeding, ignoring.")
+
+        # Reset environments and record observations
+        obs = []
+        for env in self._envs:
+            env_obs = env.reset(**kwargs)
+            env_obs = np.stack(list(env_obs.values()), axis=0)
+            obs.append(env_obs)
+        obs = np.stack(obs, axis=0)
+
+        return obs, {}
+
+    def step(self, actions: np.ndarray) -> tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, dict[str, Any]]:
+        """Take a step in the environments.
+
+        :param actions: An array of actions to take in the environments, with shape (num_envs, num_agents).
+        :type actions: np.ndarray
+        :return: A tuple containing the next observations, rewards, terminations, truncations, and an empty info dictionary.
+        :rtype: tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, dict[str, Any]]
+
+        """
+        # Deconstruct actions into list of dicts
+        actions = self._construct_actions(actions)
+
+        # Step environments and record results
+        obs, rewards, terminations, truncations = [], [], [], []
+        for env, env_actions in zip(self._envs, actions):
+            env_obs, env_rewards, env_terminations, env_truncations = env.step(env_actions)
+            env_obs = np.stack(list(env_obs.values()), axis=0)
+
+            # Record
+            obs.append(env_obs)
+            rewards.append(env_rewards)
+            terminations.append(env_terminations)
+            truncations.append(env_truncations)
+
+        # Stack results
+        obs = np.stack(obs, axis=0)
+        rewards = np.stack(rewards, axis=0)
+        terminations = np.stack(terminations, axis=0)
+        truncations = np.stack(truncations, axis=0)
+
+        return obs, rewards, terminations, truncations, {}
+
+    def render(self) -> np.ndarray:
+        """Render the environments and return a frame, currently unsupported.
+
+        :return: An array of rendered frames, with shape (num_envs, height, width, channels).
+        :rtype: np.ndarray
+
+        """
+        # Download from https://github.com/VirxEC/rlviser/releases/tag/v0.9.1, and install virtual display (Xvfb) if running headless
+        # Initialize virtual display if running headless
+        # if os.environ.get('DISPLAY') is None:
+        #     import pyvirtualdisplay
+        #     pyvirtualdisplay.Display(visible=True, size=(1280, 720), backend='xvfb').start()
+
+        # # Take screenshot
+        # def screenshot() -> np.ndarray:
+        #     import mss
+        #     with mss.mss() as sct:
+        #         frame = np.array(sct.grab(sct.monitors[0]))
+        #         frame = frame[:, :, [2, 1, 0]]  # Convert from BGRA to RGB
+        #     return frame
+        # NOTE: Rendering currently does not work
+        # TODO: Implement in RLBot and just record manually
+        raise NotImplementedError("Rendering is currently not implemented for RLGym environments. This is due to issues with RLViser.")
+
+        # Gather frames from each environment
+        frames = []
+        for env in self._envs:
+            env_frame = env.render()
+            frames.append(env_frame)
+
+        # Stack frames and return
+        return np.stack(frames, axis=0)
+
+    def copy(self, **kwargs: dict[str, Any]) -> 'VectorizedRLGymEnvironment':
+        """Copy the environment, overriding parameters in `kwargs`.
+
+        :param kwargs: Additional keyword arguments for the new environment instance.
+        :type kwargs: dict[str, Any]
+        :return: A copy of the environment with overridden parameters.
+        :rtype: VectorizedRLGymEnvironment
+
+        """
+        new_kwargs = {
+            'num_envs': self._num_envs,
+            'allow_rendering': self._allow_rendering,
+            **self._init_kwargs}
+        new_kwargs.update(kwargs)
+        return VectorizedRLGymEnvironment(**new_kwargs)
+
+
 class ENVIRONMENT_IDENTIFIERS(enum.Enum, metaclass=frl_utilities.CaseInsensitiveEnumMeta):
     """String identifiers for environment definitions, mapped to their corresponding classes."""
     GYMNASIUM = VectorizedGymEnvironment
+    RLGYM = VectorizedRLGymEnvironment